fix bug

+++ ttsd-message/src/main/java/com/tuotiansudai/message/util/UserMessageEventGenerator.java
@@ -183,7 +183,7 @@ public class UserMessageEventGenerator {
     public void generateTransferSuccessEvent(long investId) {
         Map<String, Object> invest = userMessageMetaMapper.findTransferApplicationByInvestId(investId);
         String name = (String) invest.get("name");
-        long amount = (long) invest.get("transfer_amount");
+        long amount = ((BigInteger) invest.get("transfer_amount")).longValue();
         String loginName = (String) invest.get("login_name");
 
         MessageModel messageModel = messageMapper.findActiveByEventType(MessageEventType.TRANSFER_SUCCESS);
